recursively add tags

+++ src/components/ui/TagsSidebarView/TagsSidebarView.jsx
@@ -8,6 +8,8 @@ import "./TagsSidebarView.css";
 import { setNewTagFormToggled } from "../../../redux/sidebar";
 import ColorPicker from "../ColorPicker/ColorPicker";
 import Tag from "../Tag/Tag";
+import { useEffect } from "react";
+import { throwResponseStatusError } from "../../../utils/throwResponseStatusError";
 
 const TagsSidebarView = () => {
   const tags = useSelector((state) => state.tags);
@@ -16,24 +18,43 @@ const TagsSidebarView = () => {
   const colorPickerMenuRef = useRef(null);
   const dispatch = useDispatch();
   const [tagSearchValue, setTagSearchValue] = useState("");
-  const [newTagColor, setNewTagColor] = useState("#008080");
+  const [newTagColor, setNewTagColor] = useState(null);
   const [newTagName, setNewTagName] = useState("");
-  const [updatedTagColor, setUpdatedTagColor] = useState(tags.selected?.COLOR || "");
+  const [updatedTagColor, setUpdatedTagColor] = useState(
+    determineDefaultColor(tags.selected)
+  );
   const [updatedTagName, setUpdatedTagName] = useState(tags.selected?.NAME || "");
   const [deleteWarningToggled, setDeleteWarningToggled] = useState(false);
+
+  function determineDefaultColor(tag) {
+    let defaultColor = "";
+
+    if (tag) {
+      if (tag.HAS_DEFAULT_COLOR) {
+        defaultColor = tags.colorOptions.default.find(
+          (color) => color.ID === tag.COLOR_ID
+        );
+      } else {
+        defaultColor = tags.colorOptions.userCreated.find(
+          (color) => color.ID === tag.COLOR_ID
+        );
+      }
+    }
+    return defaultColor;
+  }
 
   function handleTagClick(e, tag) {
     dispatch(selectTag(tag));
     setUpdatedTagName(tag.NAME);
-    setUpdatedTagColor(tag.COLOR);
+    setUpdatedTagColor(determineDefaultColor(tag));
   }
 
   function handleUpdatedColorChange(colorCode) {
     setUpdatedTagColor(colorCode);
   }
 
-  function handleNewColorChange(colorCode) {
-    setNewTagColor(colorCode);
+  function handleNewColorChange(color) {
+    setNewTagColor(color);
   }
 
   function handleSubmit(e) {
@@ -67,21 +88,25 @@ const TagsSidebarView = () => {
         },
         body: JSON.stringify(payload),
       });
+
+      if (response.status !== 200) throwResponseStatusError(response, "POST");
+
       const data = await response.json();
-      console.log(data);
+
+      if (!data) throw "There was an issue parsing /tags/edit response";
 
       dispatch(editTag(payload));
       dispatch(deselectTag());
-    } catch (err) {
-      console.log(err);
+    } catch (error) {
+      console.log(error);
     }
   }
 
-  async function handleNewTag(e, newTagName, newTagColor, tagId) {
+  async function handleNewTag(e, newTagName, newColor, tagId) {
     try {
       const payload = {
         name: newTagName,
-        color: newTagColor,
+        color: newColor,
         id: tagId,
       };
 
@@ -93,7 +118,12 @@ const TagsSidebarView = () => {
         },
         body: JSON.stringify(payload),
       });
+
+      if (response.status !== 200) throwResponseStatusError(response, "POST");
+
       const data = await response.json();
+
+      if (!data) throw "There was an issue parsing /tags/new response";
 
       const justCreatedTag = data.justCreatedTag;
 
@@ -102,8 +132,8 @@ const TagsSidebarView = () => {
       setNewTagColor("#008080");
       setNewTagName("");
       dispatch(setNewTagFormToggled(false));
-    } catch (err) {
-      console.log(err);
+    } catch (error) {
+      console.log(error);
     }
   }
 
@@ -121,14 +151,18 @@ const TagsSidebarView = () => {
         },
         body: JSON.stringify(payload),
       });
+
+      if (response.status !== 200) throwResponseStatusError(response, "POST");
+
       const data = await response.json();
-      console.log(data);
+
+      if (!data) throw "There was an issue parsing /tags/delete response";
 
       dispatch(deleteTag(payload));
       dispatch(deselectTag());
       setDeleteWarningToggled(false);
-    } catch (err) {
-      console.log(err);
+    } catch (error) {
+      console.log(error);
     }
   }
 
@@ -143,6 +177,7 @@ const TagsSidebarView = () => {
               placeholder="Search for a tag"
             />
           </form>
+          {tags.list.length === 0 && <p className="no-tags-found">No tags found</p>}
           {tags.list
             ?.filter((tag) => tag.NAME.includes(tagSearchValue))
             .map((tag, i) => {
@@ -151,11 +186,12 @@ const TagsSidebarView = () => {
                   onClick={(e) => handleTagClick(e, tag)}
                   className={`tag-button ${tag.SELECTED ? "selected" : ""}`}
                   key={i}
+                  title={`Tag: ${tag.NAME}`}
                 >
                   <span
                     className="color-span"
-                    style={{ backgroundColor: tag.COLOR }}
-                    title={`Color: ${tag.COLOR}`}
+                    style={{ backgroundColor: tag.COLOR_CODE }}
+                    title={`Color: ${tag.COLOR_CODE}`}
                   >
                     &nbsp;
                   </span>
@@ -171,13 +207,20 @@ const TagsSidebarView = () => {
             <button
               onClick={() => {
                 dispatch(setNewTagFormToggled(false));
+                setNewTagColor(null);
+                setNewTagName("");
               }}
+              title="Cancel / Discard New Tag"
             >
               Cancel
             </button>
             <button
               onClick={(e) => handleNewTag(e, newTagName, newTagColor)}
-              disabled={!isValidTagName(newTagName, false) || !isValidColor(newTagColor)}
+              disabled={
+                !isValidTagName(newTagName, false) ||
+                !isValidColor(newTagColor?.COLOR_CODE)
+              }
+              title="Done / Confirm New Tag"
             >
               Done
             </button>
@@ -187,6 +230,7 @@ const TagsSidebarView = () => {
             <div className="tag-name">
               <label>Tag Name</label>
               <input
+                placeholder='"Expenses", "High Priority", etc.'
                 value={newTagName}
                 onChange={(e) => setNewTagName(e.target.value)}
                 className={isValidTagName(newTagName, true) ? "" : "invalid"}
@@ -201,15 +245,18 @@ const TagsSidebarView = () => {
                 <ColorPicker
                   onColorSelect={handleNewColorChange}
                   selectedColor={newTagColor}
+                  showColorCode={true}
                 />
               </div>
-              {!isValidColor(newTagColor) && (
+              {newTagColor && !isValidColor(newTagColor?.COLOR_CODE) && (
                 <p className="invalid-text">Invalid color</p>
               )}
             </div>
           </form>
-          <Tag name={newTagName ? newTagName : '-'} color={newTagColor} />
-
+          <Tag
+            name={newTagName ? newTagName : "-"}
+            color={newTagColor || { COLOR_CODE: "#000000" }}
+          />
         </div>
       )}
       {tags.selected && (
@@ -220,7 +267,10 @@ const TagsSidebarView = () => {
                 setDeleteWarningToggled(false);
                 dispatch(setNewTagFormToggled(false));
                 dispatch(deselectTag());
+                setUpdatedTagColor(null);
+                setUpdatedTagName("");
               }}
+              title="Cancel / Discard Edit"
             >
               Cancel
             </button>
@@ -229,8 +279,12 @@ const TagsSidebarView = () => {
                 handleTagEdit(e, updatedTagName, updatedTagColor, tags.selected.ID)
               }
               disabled={
-                !isValidTagName(updatedTagName, false) || !isValidColor(updatedTagColor)
+                (updatedTagName === tags.selected?.NAME &&
+                  updatedTagColor?.COLOR_CODE === tags.selected?.COLOR_CODE) ||
+                !isValidTagName(updatedTagName, false) ||
+                !isValidColor(updatedTagColor?.COLOR_CODE)
               }
+              title="Done / Confirm Edit"
             >
               Done
             </button>
@@ -254,22 +308,26 @@ const TagsSidebarView = () => {
                 <ColorPicker
                   onColorSelect={handleUpdatedColorChange}
                   selectedColor={updatedTagColor}
+                  showColorCode={true}
                 />
               </div>
-              {!isValidColor(updatedTagColor) && (
+              {!isValidColor(updatedTagColor?.COLOR_CODE) && (
                 <p className="invalid-text">Invalid color</p>
               )}
             </div>
           </form>
-          <Tag name={tags.selected.NAME} color={tags.selected.COLOR} />
+          <Tag name={tags.selected.NAME} color={determineDefaultColor(tags.selected)} />
 
           <DownArrow />
           {updatedTagName !== tags.selected.NAME ||
-          updatedTagColor !== tags.selected.COLOR ? (
+          updatedTagColor?.COLOR_CODE !==
+            determineDefaultColor(tags.selected)?.COLOR_CODE ? (
             <Tag
               name={updatedTagName ? updatedTagName : "-"}
               color={
-                isValidColor(updatedTagColor) ? updatedTagColor : tags.selected.COLOR
+                isValidColor(updatedTagColor.COLOR_CODE)
+                  ? updatedTagColor
+                  : determineDefaultColor(tags.selected)
               }
             />
           ) : (
